requirements.txt
torch<=2.0.1
scipy
packaging
sentencepiece
datasets
deepspeed>=0.10.0
accelerate>=0.21.0,<0.23.0  # 0.23.0 will cause an incorrect learning rate schedule when using deepspeed, which is likely caused by https://github.com/huggingface/accelerate/commit/727d624322c67db66a43c559d8c86414d5ffb537
peft>=0.4.0
bitsandbytes>=0.41.1
evaluate>=0.4.0
tokenizers>=0.13.3
protobuf
# Transformers library (v4.34.0) still has a bug for left padding, 
# and significantly affect the inference and thus our evaluation performance (e.g., MMLU and TruthfulQA).
# Follwing PR is a temporary fix for it but has not been merged yet.
# See https://github.com/huggingface/transformers/pull/25284
# But this PR is not compatible with the latest version of Transformers library (v4.34.0).
# To incorporate it, we forked the Transformers library and made some changes to make it compatible with the latest version.
git+https://github.com/yizhongw/transformers.git@left_padding
openai>=1.5.0
tiktoken
rouge_score
tensorboard
wandb
gradio==3.50.2
termcolor
jsonlines
unidic-lite
einops
flash-attn==2.2.2
auto-gptq
fire
# Pradeep's fork of alpaca eval with bug fixes and better logging
# for human eval web app
flask
vllm
# vllm==0.4.0
openpyxl
prometheus-eval
bert-score